refactor: Renamed DrawModeSelector to DrawModeDropdown
